requirements.txt
amqp==5.1.1
billiard==4.1.0
blinker==1.6.2
camelot-py==0.9.0
celery==5.3.4
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
distro==1.8.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.2
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pdfminer.six==20221105
Pillow==10.0.0
prompt-toolkit==3.0.39
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.23.2
PyMuPDFb==1.23.0
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyspellchecker==0.7.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
rubicon-objc==0.4.6
six==1.16.0
tabula-py==2.7.0
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
zipp==3.16.2
